Quantization Code Move (#570)
* Moved:
- `modifier_quantization.py` to `sparseml.pytrorch.sparsification.quantization`
- Updated test location and usages
_ Style

* fix imports

* Style

* Moved `sparseml/pytorch/utils/quantization` to `sparseml/pytorch/sparsification/quantization/utils`

* Moved `sparseml/pytorch/utils/quantization` to `sparseml/pytorch/sparsification/quantization/utils`

* Style

* Style

* fix imports

* Style

* Review Comments

* fix quality + failing tests + cyclic dependency

* quality
