+++ src/sparseml/pytorch/optim/__init__.py
@@ -34,7 +34,6 @@ from .modifier_epoch import *
 from .modifier_lr import *
 from .modifier_params import *
 from .modifier_pruning import *
-from .modifier_quantization import *
 from .modifier_regularizer import *
 from .optimizer import *
 from .sensitivity_as import *
