+++ src/sparseml/pytorch/sparsification/quantization/modifier_quantization.py
@@ -46,8 +46,7 @@ except Exception:
 
 from sparseml.optim import BaseModifier, ModifierProp
 from sparseml.pytorch.optim.modifier import PyTorchModifierYAML, ScheduledModifier
-from sparseml.pytorch.utils import BaseLogger, tensors_module_forward, tensors_to_device
-from sparseml.pytorch.utils.quantization import (
+from sparseml.pytorch.sparsification.quantization.helpers import (
     add_quant_dequant,
     configure_module_default_qconfigs,
     configure_module_qat_wrappers,
@@ -57,6 +56,7 @@ from sparseml.pytorch.utils.quantization import (
     prepare_embeddings_qat,
     remove_activation_qat_by_layer_name,
 )
+from sparseml.pytorch.utils import BaseLogger, tensors_module_forward, tensors_to_device
 from sparseml.sparsification import SparsificationTypes
