+++ tests/sparseml/pytorch/sparsification/quantization/__init__.py
@@ -11,12 +11,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-"""
-Tools for quantizing and exporting PyTorch models
-"""
-
-# flake8: noqa
-
-from .helpers import *
-from .quantize_qat_export import *
